+++ js/Demo/demoMod.js
@@ -12,7 +12,7 @@ let modInfo = {
 
 // Set your version in num and name
 let VERSION = {
-	num: "2.2.2",
+	num: "2.2.3",
 	name: "Uprooted",
 }
 
@@ -64,4 +64,9 @@ function isEndgame() {
 // You can change this if you have things that can be messed up by long tick lengths
 function maxTickLength() {
 	return(3600000) // Default is 1 hour which is just arbitrarily large
+}
+
+// Use this if you need to undo inflation from an older version. If the version is older than the version that fixed the issue,
+// you can cap their current resources with this.
+function fixOldSave(oldVersion){
 }