+++ js/mod.js
@@ -60,3 +60,8 @@ function isEndgame() {
 function maxTickLength() {
 	return(3600000) // Default is 1 hour which is just arbitrarily large
 }
+
+// Use this if you need to undo inflation from an older version. If the version is older than the version that fixed the issue,
+// you can cap their current resources with this.
+function fixOldSave(oldVersion){
+}