+++ js/utils.js
@@ -301,6 +301,7 @@ function importSave(imported=undefined, forced=false) {
 		player = tempPlr;
 		player.versionType = modInfo.id
 		fixSave()	
+		versionCheck()
 		save()
 		window.location.reload()
 	} catch(e) {
@@ -317,7 +318,10 @@ function versionCheck() {
 	}
 	
 	if (setVersion) {
-		if (player.versionType == modInfo.id && VERSION.num > player.version) player.keepGoing = false
+		if (player.versionType == modInfo.id && VERSION.num > player.version) {
+			player.keepGoing = false
+			if (fixOldSave) fixOldSave(player.version)
+		} 
 		player.versionType = getStartPlayer().versionType
 		player.version = VERSION.num
 		player.beta = VERSION.beta